feat: allow json string as input
